axie-scholar-utilities/source/pyproject.toml
[tool.poetry]
name = "Axie Scholar Utilities"
version = "3.0.0"
description = "Package to automate activites related to manage Axie Infinity Scholars. Specially aimed at managers with large scholar roasters."
authors = ["Ferran Marin <ferran.marin.llobet@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
docopt = "^0.6.2"
web3 = "^5.28.0"
jsonschema = "^3.2.0"
qrcode = "^7.3.1"
Pillow = "^9.0.1"
trezor = "^0.13.0"
axie-utils = "^2.0.4"

[tool.poetry.dev-dependencies]
mock = "^4.0.3"
pytest-mock = "^3.7.0"
pytest-asyncio = "^0.18.3"
requests-mock = "^1.9.3"
freezegun = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
